backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
pdfplumber>=0.11.0
astrapy>=1.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-httpx>=0.30.0
ortools>=9.9
